Add mouse navigation and improve performance

+++ src/mancet.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 typedef struct {
-	float x, y;
+	double x, y;
 } vec2;
 
 void mancet_run(uint32_t window_width, uint32_t window_height, uint32_t pixel_width);
